:recycle: Extract walkToFrame
